% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
<<<<<<< HEAD
=======
\encoding{utf-8}
>>>>>>> bf9d46bd1ca6df79852129e9ad18df873cc0b89a
\name{dexclude}
\alias{dexclude}
\title{Exclude declarations from set}
\usage{
dexclude(decls1, decls2)
}
\arguments{
\item{decls1}{First set of declarations}

<<<<<<< HEAD
\item{decls1}{The second set of declarations}
=======
\item{decls2}{The second set of declarations, or their ids.}
>>>>>>> bf9d46bd1ca6df79852129e9ad18df873cc0b89a
}
\description{
Excludes declarations from set that belongs to other set
}
\examples{
library(dplyr)
mps2016 <- 
   download_declarations("народний депутат", declaration_year = "2016") \%>\% 
   dexclude(download_declarations("помічник народного депутата", declaration_year = "2016"))
}
\keyword{dexclude}
