% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{dintersect}
\alias{dintersect}
\title{Declarations intersection}
\usage{
dintersect(decls1, decls2)
}
\arguments{
\item{decls1}{Declarations set}

\item{decls2}{Declarations set}
}
\description{
Function to find declarations within other set of declarations (intersect two sets)
}
\examples{
library(dplyr)
mps2016_bitcoins <- 
    download_declarations("народний депутат", declaration_year = "2016") \%>\% 
    dintersect(download_declarations("біткойн", deepseach = TRUE))
}
\keyword{dintersect}
