% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{download_declarations}
\alias{download_declarations}
\title{Download declarations}
\usage{
download_declarations(q = NULL, deepsearch = FALSE,
  declaration_year = NULL, doc_type = NULL, post_type = NULL,
  region_type = NULL, region_value = NULL)
}
\arguments{
\item{q}{Search query}

\item{deepsearch}{Should website search in all declarations field, not only in name and workpost?}

\item{declaration_year}{Character. The year of declaration.}

\item{doc_type}{Character. The type of declaration.}

\item{post_type}{Character. The type of post}

\item{region_type}{Should it be search in regions where declarer is registered, where he lives or where has realty?}

\item{region_value}{Region query value.}
}
\description{
Downloads declarations from declarations.com.ua
}
\examples{
library(dplyr)
poroshenko2016 <- 
    download_declarations("порошенко петро олексійович", declaration_year = "2016") \%>\% 

}
\keyword{download_declarations}
