% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{extract_names}
\alias{extract_names}
\title{Exctract names}
\usage{
extract_names(decls)
}
\arguments{
\item{decls}{The declarations set}
}
\description{
Extracts names of the declarers
}
\examples{
library(dplyr)
bitcoins2016 <- 
    download_declarations("біткойн", declaration_year = "2016", deepsearch = T) \%>\% 
    extract_names()
}
\keyword{extract_names}
