% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\encoding{utf-8}
\name{get_corrected}
\alias{get_corrected}
\title{Gets only corrected versions of declarations}
\usage{
get_corrected(decls)
}
\arguments{
\item{decls}{Declarations set}
}
\description{
Exclude from a set declarations that have corrected version.
}
\examples{
library(dplyr)
zalishchuk2015_corrected <- 
   download_declarations("заліщук світлана петрівна", declaration_year = "2015") \%>\% 
   get_corrected()
}
\keyword{get_corrected}
