% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\encoding{utf-8}
\name{related_companies}
\alias{related_companies}
\title{Related companies}
\usage{
related_companies(decls)
}
\description{
Function to find companies related to declarers
}
\examples{
library(dplyr)
poroshenko_companies <- 
    download_declarations("порошенко петро олексійович", declaration_year = "2016") \%>\% 
    related_companies()
}
\keyword{related_companies}
