% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{step_to_df}
\alias{step_to_df}
\title{Extract declarations sections}
\usage{
step_to_df(decls, step, add_rights = FALSE, guarantor = FALSE,
  guarantor_realty = FALSE)
}
\arguments{
\item{step}{Step (section) of declarations to be extracted. Works correctly if receives values between 2 and 16}

\item{add_rights}{Logical. Should information on additional (not belonged to declarer or his family members) rights be saved? Default to FALSE.}

\item{guarantor_realty}{Logical. Should information on loan guarantors' realty be saved? Defaults to FALSE.}

\item{d}{Declarations set}

\item{guarantor.}{Logical. Should information on loan guarantors be saved? Defaults to FALSE.}
}
\description{
Extract specific sections of declarations into dataframe. For now - ignoring "rights" and "guarantor" subsections
}
\details{
The value is always list of 4 data frames. The data frames are available by names "data", "add_rights", "guarantor", "guarantor_realty". If corresponded parameters equal FALSE, these dataframes are always blank. If these parameters are set to TRUE, the function works slower.
}
\examples{
library(dplyr)
mps2016_realty <- 
   download_declarations("народний депутат", declaration_year = "2016") \%>\% 
   step_to_df(3)$data
}
\keyword{step_to_df}
