% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{dexclude}
\alias{dexclude}
\title{Exclude declarations from set}
\usage{
dexclude(decls1, decls2)
}
\arguments{
\item{decls1}{First set of declarations}

\item{decls2}{The second set of declarations, or their ids.}
}
\description{
Excludes declarations from set that belongs to other set
}
\examples{
library(dplyr)
mps2016 <- 
   download_declarations("народний депутат", declaration_year = "2016") \%>\% 
   dexclude(download_declarations("помічник народного депутата", declaration_year = "2016"))
}
\keyword{dexclude}
