% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\encoding{utf-8}
\name{download_declarations}
\alias{download_declarations}
\title{Download declarations}
\usage{
download_declarations(q = NULL, deepsearch = FALSE,
  declaration_year = NULL, doc_type = NULL, post_type = NULL,
  region_type = NULL, region_value = NULL)
}
\arguments{
\item{q}{Search query}

\item{deepsearch}{Should website search in all declarations field, not only in name and workpost?}

\item{declaration_year}{Character or numeric. The year of declaration.}

\item{doc_type}{Character or numeric. The type of declaration. 1 - "Щорічна", 2 - "Перед звільненням", 3 - "Після звільнення", 4 - "Кандидата на посаду", 5 - "Форма змін".}

\item{post_type}{Character or numeric. The type of declarer's position. Accepting vector longer than 1 element. 1 or "державної" for state authorities position, 2 or "місцевого" for local authorities, 3 or "юридичної" for state-owned enterprises.}

\item{region_type}{Should it be search in regions where declarer is registered (1 or "region"), where he lives (2 or "actual_region") or where owns realty (3 or "estate_region")?}

\item{region_value}{Region query value. Substring that can identify region name.}
}
\description{
Downloads declarations from declarations.com.ua
}
\examples{
library(dplyr)
poroshenko2016 <- 
    download_declarations("порошенко петро олексійович", declaration_year = 2016, declaration_type = 1)

}
\keyword{download_declarations}
