% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecl.R
\name{step_to_df}
\alias{step_to_df}
\title{Extract declarations sections}
\usage{
step_to_df(decls, step)
}
\arguments{
\item{step}{Step (section) of declarations to be extracted. Works correctly if receives values between "step_2" and "step_15"}

\item{d}{Declarations set}
}
\description{
Extract specific sections of declarations into dataframe. For now - ignoring "rights" and "guarantor" subsections
}
\examples{
library(dplyr)
mps2016_realty <- 
   download_declarations("народний депутат", declaration_year = "2016") \%>\% 
   step_to_df("step_3")
}
\keyword{step_to_df}
